Refactor: Update FastAPI, fix DB connection, improve tests
